PromQL: lay out the vector match join in a builder
Move the join layout out of the translator into
VectorBinaryOperatorLayout, a fluent builder that owns the three
rules of the block: how the sides are ordered (probe vs build), how
the fields are placed (packed match keys, carried build columns,
null-filled labels) and what is projected. The translator translates
the operands and feeds them in.

Assert the Filter.NONE case explicitly: without on/ignoring the
verifier only admits operands with concrete label sets, so the
header needs no widening.
